+++ .github/workflows/validate.yml
@@ -1,4 +1,4 @@
-name: Validate gauges, tokens, and validators
+name: Validate vaults, tokens, and validators
 
 on:
   pull_request:
