+++ scripts/validate-vaults.ts
@@ -2,19 +2,19 @@ import { readdirSync } from 'node:fs'
 import { createPublicClient, http } from 'viem'
 
 import { supportedChains } from '@/config/chains'
-import type { GaugesSchema } from '@/types/gauges'
+import type { GaugesSchema } from '@/types/vaults'
 
 import { getFile } from './_/get-file'
 import { getJsonFile } from './_/get-json-file'
 import { isValidNetwork } from './_/is-valid-network'
 import { outputScriptStatus } from './_/output-script-status'
-import { validateGaugeDetails } from './_/validate-gauge-details'
 import { validateList } from './_/validate-list'
+import { validateVaultDetails } from './_/validate-vault-details'
 
-const schema = getFile('schema/gauges-schema.json')
-const folderPath = 'src/gauges'
+const schema = getFile('schema/vaults-schema.json')
+const folderPath = 'src/vaults'
 
-const validateGauges = async ({
+const validateVaults = async ({
   network,
 }: {
   network: keyof typeof supportedChains
@@ -33,7 +33,7 @@ const validateGauges = async ({
   })
 
   validateList({ errors, list: gauges, schema, type: 'gauges' })
-  await validateGaugeDetails({
+  await validateVaultDetails({
     errors,
     gauges: gauges.gauges,
     network,
@@ -49,7 +49,7 @@ readdirSync(folderPath).forEach(async (file) => {
     throw new Error(`Unsupported network: ${network}`)
   }
 
-  await validateGauges({
+  await validateVaults({
     network,
   })
 })
